+++ src/main/java/com/devbliss/gpullr/domain/PullRequest.java
@@ -40,9 +40,12 @@ public class PullRequest {
 
   public String title;
 
+  @Column(unique = true)
   public String url;
 
   public ZonedDateTime createdAt;
+  
+  public ZonedDateTime closedAt;
 
   @NotNull
   @Enumerated(EnumType.STRING)