+++ src/main/java/com/devbliss/gpullr/exception/BadRequestException.java
@@ -1,0 +1,21 @@
+package com.devbliss.gpullr.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+/**
+ * To be thrown whenever a user or the frontend has sent invalid or incomplete data or requests.
+ */
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class BadRequestException extends RuntimeException {
+
+  private static final long serialVersionUID = -1749257237661508754L;
+
+  public BadRequestException(String message) {
+    super(message);
+  }
+
+  public BadRequestException(Throwable cause) {
+    super(cause);
+  }
+}