+++ src/main/java/com/devbliss/gpullr/repository/UserRepository.java
@@ -2,16 +2,17 @@ package com.devbliss.gpullr.repository;
 
 import com.devbliss.gpullr.domain.User;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Component;
 
 /**
  * Persistence layer for {@link com.devbliss.gpullr.domain.User} objects.
  */
-@Component
 public interface UserRepository extends CrudRepository<User, Integer> {
 
   List<User> findAll();
 
   List<User> findByCanLoginIsTrue();
+  
+  Optional<User> findById(Integer id);
 }