+++ src/main/java/com/devbliss/gpullr/util/TaskSchedulerProducer.java
@@ -1,7 +1,9 @@
 package com.devbliss.gpullr.util;
 
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 
@@ -13,10 +15,12 @@ import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
  */
 @Configuration
 public class TaskSchedulerProducer {
-  
+
   @Bean
+  @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
   public TaskScheduler createThreadPoolTaskScheduler() {
     ThreadPoolTaskScheduler executor = new ThreadPoolTaskScheduler();
+    executor.setPoolSize(5);
     executor.initialize();
     return executor;
   }